Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -3,7 +3,7 @@ class CreateCostumes < ActiveRecord::Migration[5.1]
   def change
     create_table :costumes do |c|
       c.string :name
-      c.integer :price
+      c.float :price
       c.string :size
       c.string :image_url
       c.timestamps
